修复了删除的商品没有评论返回值错误的bug

+++ src/com/hamburgerking/service/impl/CommentsServiceImpl.java
@@ -38,6 +38,8 @@ public class CommentsServiceImpl implements CommentsService {
      * @return 是否删除成功
      */
     public boolean deleteGoodOfComments(String gid){
+        if (commentsDAO.checkGoodOfComment(Integer.parseInt(gid))==0)
+            return true;
         return commentsDAO.deleteGoodOfComments(Integer.parseInt(gid));
     }
 }
